rename package pastie->pystemon
