+++ pystemon/sites/pastieberylia.py
@@ -1,13 +1,12 @@
 import json
-from pastie.pastie import Pastie
+from pystemon.pastie import Pastie
 import logging.handlers
 logger = logging.getLogger('pystemon')
 
-class PastieCdvLt(Pastie):
+class PastieBerylia(Pastie):
     '''
-    Custom Pastie class for the cdv.lt site
-    This class overloads the fetch_pastie function to do the form submit
-    to get the raw pastie
+    Custom Pastie class for the berylia.org site, related to the LockedShields cyber exercise
+    This class overloads the fetch_pastie function to extract the pastie from the page
     '''
 
     def __init__(self, site, pastie_id):
@@ -21,6 +20,7 @@ class PastieCdvLt(Pastie):
             json_pastie = json.loads(downloaded_page)
             if json_pastie:
                 # and extract the code
-                self.pastie_content = json_pastie['snippet']['snippetData']
+                self.pastie_content = json_pastie['paste']
         return self.pastie_content
+
